Stockage config et operation sur internet

+++ src/app/dbtable/dbtable.component.ts
@@ -10,24 +10,59 @@ import {environment} from "../../environments/environment";
 export class DbtableComponent implements OnInit {
 
   @Input("table") table:string="";
+  @Input() source:any="local";
   @Input("max_len") max_len=20;
   @Input("excludes") excludes:string="";
   rows:any[]=[];
-  cols: string[]=[];
+  cols:string[]=[];
+  @Input() showClear=false;
+  @Input() dictionnary:any={};
 
   constructor(
     public httpClient:HttpClient
   ) { }
 
-  refresh(){
-    this.httpClient.get(environment.server+"/api/tables/"+this.table+"?excludes="+this.excludes).subscribe((rows:any)=>{
-      this.rows=rows;
-      if(this.cols.length==0){
-        for(let k in this.rows[0]){
-          this.cols.push(k);
+  update_cols(cols:string[]){
+    let rc=[];
+    for(let c of cols){
+      if(!this.dictionnary.hasOwnProperty(c))this.dictionnary[c]=c;
+      rc.push(this.dictionnary[c]);
+    }
+    return rc;
+  }
+
+  update_cells(rows:any[]){
+    for(let i=0;i<rows.length;i++){
+      for(let k of Object.keys(rows[i])){
+        let v=rows[i][k];
+        if(typeof(v)=="boolean"){
+          v=v ? "X" : "";
         }
+        rows[i][this.dictionnary[k]]=v;
       }
-    })
+    }
+    return(rows);
+  }
+
+
+
+  refresh(){
+    if(this.source=="db"){
+      this.httpClient.get(environment.server+"/api/tables/"+this.table+"?excludes="+this.excludes).subscribe((rows:any)=>{
+        this.rows=rows;
+        if(this.cols.length==0){
+          for(let k in this.rows[0]){
+            this.cols.push(k);
+          }
+        }
+      })
+    } else {
+      if(this.source){
+        this.cols=this.update_cols(Object.keys(this.source[0]));
+        this.rows=this.update_cells(this.source);
+      }
+    }
+
   }
 
   ngOnInit(): void {
@@ -35,9 +70,11 @@ export class DbtableComponent implements OnInit {
   }
 
   clear() {
-    this.httpClient.delete(environment.server+"/api/tables/"+this.table).subscribe(()=>{
-      this.refresh();
-    })
+    if(this.source!="local"){
+      this.httpClient.delete(environment.server+"/api/tables/"+this.table).subscribe(()=>{
+        this.refresh();
+      })
+    }
   }
 
   truncate(txt:string) {
